+++ packages/isaacscript-common/src/features/customStage/shadows.ts
@@ -4,7 +4,7 @@ import { getRandomArrayElement } from "../../functions/array";
 import { spawnEffectWithSeed } from "../../functions/entitiesSpecific";
 import { removeCharactersBefore } from "../../functions/string";
 import { CustomStage } from "../../interfaces/private/CustomStage";
-import { ISAACSCRIPT_CUSTOM_STAGE_GFX_PATH } from "./customStageConstants";
+import { ISAACSCRIPT_CUSTOM_STAGE_GFX_PATH } from "./constants";
 import v from "./v";
 
 type ShadowAnimation = "1x1" | "1x2" | "2x1" | "2x2";