+++ packages/isaacscript-common/src/features/saveDataManager/exports.ts
@@ -4,6 +4,7 @@ import { errorIfFeaturesNotInitialized } from "../../featuresInitialized";
 import { deepCopy } from "../../functions/deepCopy";
 import { isString } from "../../functions/types";
 import { SaveData } from "../../interfaces/SaveData";
+import { SAVE_DATA_MANAGER_FEATURE_NAME } from "./constants";
 import {
   forceSaveDataManagerLoad,
   forceSaveDataManagerSave,
@@ -14,7 +15,6 @@ import {
   saveDataDefaultsMap,
   saveDataMap,
 } from "./maps";
-import { SAVE_DATA_MANAGER_FEATURE_NAME } from "./saveDataManagerConstants";
 
 /**
  * This is the entry point to the save data manager, a system which provides two major features: