+++ packages/docs/docs/main/change-log.md
@@ -11,6 +11,7 @@ This page lists the changes to the IsaacScript framework.
 
 ## September 20th, 2022 (Unreleased)
 
+- The save data manager will now restore any data on a `run` or `level` object when the Glowing Hour Glass is used to what it was when the room was entered.
 - Added the following helper functions:
   - `isTearFromPlayer`
   - `isTearFromFamiliar`